requirements.txt
numpy
pandas
matplotlib
scipy
typer
PyYAML
streamlit
